+++ RELEASENOTES.md
@@ -17,6 +17,9 @@
     *   Use `SingleThreadExecutor` for releasing `AudioTrack` instances to avoid
         OutOfMemory errors when releasing multiple players at the same time
         ([#10057](https://github.com/google/ExoPlayer/issues/10057)).
+    *   Fix implementation of `setDataSourceFactory` in
+        `DefaultMediaSourceFactory`, which was non-functional in some cases
+        ([#116](https://github.com/androidx/media/issues/116)).
 *   Extractors:
     *   Add support for AVI
         ([#2092](https://github.com/google/ExoPlayer/issues/2092)).